requirements.txt
fastapi
uvicorn[standard]
pandas
openpyxl
PyPDF2
python-docx
dnspython
tldextract
requests
python-multipart
pydantic
redis
rq
aiodns
tqdm